+++ core/resources/homepage/tpls/hp.latestversion.php
@@ -5,15 +5,13 @@ $neardLatestVersion =  Util::getLatestVersion();
 
 if ($neardLatestVersion != null && version_compare($neardCurrentVersion, $neardLatestVersion, '<')) {
     $fullVersionUrl = Util::getVersionUrl($neardLatestVersion);
-    $patchVersionUrl = Util::getPatchUrl($neardCurrentVersion, $neardLatestVersion);
 ?>
 
 <div class="alert alert-dismissable alert-success">
     <button data-dismiss="alert" class="close" type="button">×</button>
     <h4><?php echo $neardLang->getValue(Lang::CHECK_VERSION_AVAILABLE_TEXT); ?></h4>
     <p>
-      <a role="button" class="btn btn-success" href="<?php echo $patchVersionUrl; ?>" target="_blank"><i class="fa fa-download"></i> <?php echo $neardLang->getValue(Lang::DOWNLOAD); ?> <strong>Neard <?php echo $neardCurrentVersion; ?>-<?php echo $neardLatestVersion; ?> Patch</strong><br /><small>neard-<?php echo $neardCurrentVersion; ?>-<?php echo $neardLatestVersion; ?>.zip (<?php echo Util::getRemoteFilesize($patchVersionUrl); ?>)</small></a>
-      <a role="button" class="btn btn-success" href="<?php echo $fullVersionUrl; ?>" target="_blank"><i class="fa fa-download"></i> <?php echo $neardLang->getValue(Lang::DOWNLOAD); ?> <strong>Neard <?php echo $neardLatestVersion; ?> Full</strong><br /><small>neard-<?php echo $neardLatestVersion; ?>.zip (<?php echo Util::getRemoteFilesize($fullVersionUrl); ?>)</small></a>
+      <a role="button" class="btn btn-success" href="<?php echo $fullVersionUrl; ?>" target="_blank"><i class="fa fa-download"></i> <?php echo $neardLang->getValue(Lang::DOWNLOAD); ?> <strong>Neard <?php echo $neardLatestVersion; ?></strong><br /><small>neard-<?php echo $neardLatestVersion; ?>.zip (<?php echo Util::getRemoteFilesize($fullVersionUrl); ?>)</small></a>
     </p>
     <p><?php echo sprintf($neardLang->getValue(Lang::READ_CHANGELOG), '<a href="#" data-toggle="modal" data-target=".modal-changelog">', '</a>'); ?></p>
 </div>